Extract container and provider
