+++ packages/docusaurus-theme-live-codeblock/src/theme/Playground/index.tsx
@@ -6,63 +6,11 @@
  */
 
 import React, {type ReactNode} from 'react';
-import {LiveProvider} from 'react-live';
-import {usePrismTheme, useThemeConfig} from '@docusaurus/theme-common';
-import PlaygroundPreview from '@theme/Playground/Preview';
-import PlaygroundEditor from '@theme/Playground/Editor';
+import PlaygroundProvider from '@theme/Playground/Provider';
+import PlaygroundContainer from '@theme/Playground/Container';
+import PlaygroundLayout from '@theme/Playground/Layout';
 
 import type {Props} from '@theme/Playground';
-import type {ThemeConfig} from '@docusaurus/theme-live-codeblock';
-
-import styles from './styles.module.css';
-
-// this should rather be a stable function
-// see https://github.com/facebook/docusaurus/issues/9630#issuecomment-1855682643
-const DEFAULT_TRANSFORM_CODE = (code: string) => `${code};`;
-
-type PlaygroundProviderProps = Omit<Props, 'children'> & {
-  code: string | undefined;
-  children: ReactNode;
-};
-
-function PlaygroundProvider(props: PlaygroundProviderProps): ReactNode {
-  const prismTheme = usePrismTheme();
-  const noInline = props.metastring?.includes('noInline') ?? false;
-  return (
-    <LiveProvider
-      noInline={noInline}
-      theme={prismTheme}
-      {...props}
-      code={props.code?.replace(/\n$/, '')}
-      transformCode={props.transformCode ?? DEFAULT_TRANSFORM_CODE}>
-      <PlaygroundContent />
-    </LiveProvider>
-  );
-}
-
-function useLiveCodeBlockThemeConfig() {
-  const themeConfig = useThemeConfig() as unknown as ThemeConfig;
-  return themeConfig.liveCodeBlock;
-}
-
-function PlaygroundContent(): ReactNode {
-  const {playgroundPosition} = useLiveCodeBlockThemeConfig();
-  return (
-    <>
-      {playgroundPosition === 'top' ? (
-        <>
-          <PlaygroundPreview />
-          <PlaygroundEditor />
-        </>
-      ) : (
-        <>
-          <PlaygroundEditor />
-          <PlaygroundPreview />
-        </>
-      )}
-    </>
-  );
-}
 
 export default function Playground({
   children,
@@ -70,10 +18,10 @@ export default function Playground({
   ...props
 }: Props): ReactNode {
   return (
-    <div className={styles.playgroundContainer}>
+    <PlaygroundContainer>
       <PlaygroundProvider code={children} {...props}>
-        <PlaygroundContent />
+        <PlaygroundLayout />
       </PlaygroundProvider>
-    </div>
+    </PlaygroundContainer>
   );
 }
